# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  billing.CreateChargeInput:
    properties:
      amount_cents:
        type: integer
      charge_type:
        type: string
      contract_id:
        type: string
      customer_id:
        type: string
      description:
        type: string
      due_date:
        type: string
      milestone_id:
        type: string
      payment_method:
        type: string
    type: object
  billing.PaymentLinkResponse:
    properties:
      charge_id:
        type: string
      link:
        type: string
    type: object
  billing.PaymentQRResponse:
    properties:
      charge_id:
        type: string
      qr_code:
        type: string
    type: object
  contracts.CreateContractInput:
    properties:
      amount_cents:
        type: integer
      auto_renew:
        type: boolean
      billing_type:
        type: string
      contract_type:
        type: string
      customer_id:
        type: string
      deliverables:
        type: string
      duration_months:
        type: integer
      end_date:
        type: string
      payment_terms:
        type: string
      penalties:
        type: string
      sla:
        type: string
      start_date:
        type: string
      title:
        type: string
    type: object
  contracts.GenerateContractInput:
    properties:
      editable_content:
        type: string
      template_name:
        type: string
    type: object
  contracts.SendSignatureInput:
    properties:
      signer_email:
        type: string
      signer_name:
        type: string
    type: object
  customers.CreateCustomerInput:
    properties:
      address:
        type: string
      company:
        type: string
      cpf_cnpj:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      preferred_payment_method:
        type: string
    type: object
  customers.UpdateCustomerInput:
    properties:
      address:
        type: string
      company:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      preferred_payment_method:
        type: string
    type: object
...
  finance.CashFlowPoint:
    properties:
      date:
        type: string
      in_cents:
        type: integer
      pending_cents:
        type: integer
    type: object
  finance.Dashboard:
    properties:
      active_contracts:
        type: integer
      expiring_contracts:
        type: integer
      monthly_revenue_cents:
        type: integer
      mrr_cents:
        type: integer
      payments_received:
        type: integer
      pending_payments:
        type: integer
    type: object
  finance.DefaultMetrics:
    properties:
      default_rate_percent:
        type: number
      growth_percent:
        type: number
      overdue_charges:
        type: integer
    type: object
...
  problem.BaseErr:
    properties:
      details:
        description: Details contém informações adicionais opcionais sobre o erro.
      error:
        description: Title é o título legível do tipo de problema.
        type: string
      errorDescription:
        description: ErrorDescription é uma explicação específica desta ocorrência.
        type: string
      instance:
        description: Instance é o URI da requisição que originou o erro.
        type: string
      statusCode:
        description: StatusCode é o código HTTP associado ao problema.
        type: integer
      timestamp:
        description: Timestamp indica quando o erro ocorreu (UTC, ISO 8601).
        type: string
      type:
        description: Type é um URI que identifica o tipo do problema (único por categoria
          de erro).
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API modular de pagamentos com autenticação Bearer Token.
  title: AggiPay API
  version: 1.0.0
paths:
...
  /charges:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar cobranças
      tags:
      - Billing
    post:
      consumes:
      - application/json
      description: Gera cobrança única, mensal ou por milestone
      parameters:
      - description: Dados da cobrança
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/billing.CreateChargeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar cobrança
      tags:
      - Billing
  /charges/{id}/pay-link:
    post:
      parameters:
      - description: ID da cobrança
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/billing.PaymentLinkResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Obter link de pagamento
      tags:
      - Billing
  /charges/{id}/pay-qr:
    post:
      parameters:
      - description: ID da cobrança
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/billing.PaymentQRResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Obter QR Code de pagamento
      tags:
      - Billing
  /contracts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar contratos
      tags:
      - Contracts
    post:
      consumes:
      - application/json
      description: Cadastra contrato de serviço de software
      parameters:
      - description: Dados do contrato
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contracts.CreateContractInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar contrato
      tags:
      - Contracts
  /contracts/{id}:
    get:
      parameters:
      - description: ID do contrato
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Buscar contrato
      tags:
      - Contracts
  /contracts/{id}/generate:
    post:
      consumes:
      - application/json
      description: Gera artefatos PDF/editável para assinatura
      parameters:
      - description: ID do contrato
        in: path
        name: id
        required: true
        type: string
      - description: Template e conteúdo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contracts.GenerateContractInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Gerar versão de contrato
      tags:
      - Contracts
  /contracts/{id}/send-signature:
    post:
      consumes:
      - application/json
      description: Cria documento no Clicksign e retorna URL de assinatura
      parameters:
      - description: ID do contrato
        in: path
        name: id
        required: true
        type: string
      - description: Dados do signatário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contracts.SendSignatureInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Enviar para assinatura
      tags:
      - Contracts
  /customers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar clientes
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Cadastra um cliente para gestão financeira/contratos
      parameters:
      - description: Dados do cliente
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customers.CreateCustomerInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar cliente
      tags:
      - Customers
  /customers/{id}:
    get:
      description: Retorna cadastro e resumo financeiro do cliente
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Detalhar cliente
      tags:
      - Customers
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customers.UpdateCustomerInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Atualizar cliente
      tags:
      - Customers
  /finance/cashflow:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/finance.CashFlowPoint'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Fluxo de caixa
      tags:
      - Finance
  /finance/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.Dashboard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Dashboard financeiro
      tags:
      - Finance
  /finance/defaults:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.DefaultMetrics'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Métricas de inadimplência e crescimento
      tags:
      - Finance
...
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"